+++ tests/testthat/test-surv_xgboost_aft.R
@@ -126,3 +126,46 @@ test_that(desc = "test nested cv, bayesian - surv_xgboost_aft", code = {
     what = "mlexCV"
   ))
 })
+
+learner_args <- list(
+  objective = "survival:aft",
+  eval_metric = "aft-nloglik",
+  nrounds = 20L
+)
+
+test_that(desc = "test nested cv, grid - surv_xgboost_aft", code = {
+  surv_xgboost_aft_optimizer <- mlexperiments::MLNestedCV$new(
+    learner = LearnerSurvXgboostAft$new(
+      metric_optimization_higher_better = FALSE
+    ),
+    strategy = "grid",
+    fold_list = fold_list,
+    k_tuning = 3L,
+    ncores = ncores,
+    seed = seed
+  )
+
+  set.seed(seed)
+  random_grid <- sample(seq_len(nrow(param_list_xgboost)), 3)
+  surv_xgboost_aft_optimizer$parameter_grid <-
+    param_list_xgboost[random_grid, ]
+  surv_xgboost_aft_optimizer$learner_args <- learner_args
+  surv_xgboost_aft_optimizer$split_type <- "stratified"
+  surv_xgboost_aft_optimizer$split_vector <- split_vector
+
+  surv_xgboost_aft_optimizer$performance_metric <- c_index
+
+  # set data
+  surv_xgboost_aft_optimizer$set_data(
+    x = train_x,
+    y = train_y
+  )
+
+  cv_results <- surv_xgboost_aft_optimizer$execute()
+  expect_type(cv_results, "list")
+  expect_equal(dim(cv_results), c(3, 10))
+  expect_true(inherits(
+    x = surv_xgboost_aft_optimizer$results,
+    what = "mlexCV"
+  ))
+})